+++ CHANGELOG.md
@@ -8,17 +8,19 @@ Unreleased
    - `TextBoxStyleBuilder::default()`
    - `TabSize::default()`
  * [#134] `Tail` plugin
+ * [#135] Allow using the built-in plugins without the `plugin` feature.
 
 ## Changed:
 
  * **breaking** [#133] `TextBoxStyle` and `TextBoxStyleBuilder` no longer implement the `Default` trait.
 
 ## Removed:
- 
+
  * [#134] `Scrolling` vertical alignment
 
 [#133]: https://github.com/embedded-graphics/embedded-text/pull/133
 [#134]: https://github.com/embedded-graphics/embedded-text/pull/134
+[#135]: https://github.com/embedded-graphics/embedded-text/pull/135
 
 0.5.0-beta.2 (2021-07-10)
 ==========================